requirements.txt
# Core Dependencies
numpy
psutil
transformers==4.51.1
vllm==0.8.5
ray==2.43.0
xformers==0.0.29.post2
einops==0.8.1
datasets==4.4.2
pandas==2.3.3
accelerate==1.12.0
wandb==0.23.1
hydra-core==1.3.2
omegaconf==2.3.0
peft==0.14.0
tqdm==4.67.1

# Agent & Environment
google-genai
google-generativeai
boto3
gymnasium
langchain
langchain-openai
alfworld==0.4.2
textworld==1.6.2

# Utilities
aiohttp==3.13.2
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
requests==2.32.5
python-dotenv==1.2.1
click==8.3.1
rich==14.2.0
pyyaml==6.0.3
jinja2==3.1.6
